pyproject.toml
[project]
name = "polarsen"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Julien Brayere", email = "julien.brayere@polarsen.io" }]

#default-groups = "all"

[dependency-groups]
default = [
    "asyncpg>=0.30.0",
    "piou>=0.20.0",
    "niquests>=3.14.1",
    "tracktolib>=0.55.0",
    "urllib3-future>=2.13.909",
]

dev = [
    "pyright>=1.1.398",
    "pytest>=8.3.5",
    "ruff>=0.12.1",
    "datamodel-code-generator>=0.31.2",
    "httpx>=0.28.1",
    "psycopg>=3.2.9",
    "pytest-cov>=6.2.1",
    "minio>=7.2.16",
    "faker>=37.6.0",
    "tracktolib>=0.57.0",
]

llms = [
    "google-genai>=1.29.0",
    # Needed by google-genai
    "requests>=2.32.0",
    "mistralai>=1.6.0",
    "openai>=1.82.0",
    "pydantic>=2.11.2"
]

api = [
    "fastapi[standard]>=0.115.12",
    "bcrypt>=4.3.0",
    "botocore>=1.40.40",
    "sentry-sdk[fastapi]>=2.32.0",
]

telegram = [
    "python-telegram-bot>=22.1",
    "watchfiles>=1.0.5",
    "sentry-sdk>=2.32.0",
]

bump = [
    "commitizen>=4.8.3",
]

cli = [
    "botocore>=1.40.40",
    "sentry-sdk>=2.32.0",
]

[tool.pyright]
include = ["polarsen", "tests"]
exclude = []

pythonVersion = "3.12"
pythonPlatform = "Linux"

[tool.pytest.ini_options]
addopts = "--cov=polarsen -s -q -x --no-cov-on-fail"
testpaths = [
    "tests"
]
python_files = "*.py"
filterwarnings = [
    'ignore:datetime.datetime.utcnow\(\) is deprecated:DeprecationWarning',
]


[tool.ruff]
line-length = 120
target-version = "py312"


[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    # imported but unused
    "F401",
    # unable to detect undefined names
    "F403",
    #
    "F405"
]
"conftest.py" = [
    # unable to detect undefined names
    "F403",
    #
    "F405"
]


[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"


[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"


[tool.commitizen]
version = "0.1.0"
name = "cz_conventional_commits"
tag_format = "$version"
version_files = [
    "pyproject.toml:version"
]
bump_message = "chore(release): release $current_version → $new_version"


[tool.setuptools.packages.find]
include = ["polarsen*"]
exclude = ["sql*", "data*", "infra*"]
